条項2の後半を削除
Franz says:  This just clarifies what the contract already describes and is not a change in the intent of the contract. Both section 1 and 3 (included below) are very clear about "one license for one site." This extra clause was causing confusion with legal departments at larger clients on if they had to explicity ask each marketplace vendor to allow them to make modifications to the code they buy. That was never the intent, so we removed the clause to clarify.